Refactor and move account related routes
