+++ main.py
@@ -53,7 +53,6 @@ from app import create_app
 app = create_app()
 logger = app.extensions["logger"]
 utils.log_event(message="TEST")
-# region MODELS
 
 from typing import Optional
 from werkzeug.security import generate_password_hash, check_password_hash
@@ -70,165 +69,7 @@ from nfdi_search_engine.services.tracking_task_proc import TrackingTaskProcessor
 from nfdi_search_engine.web import decorators
 from nfdi_search_engine.common import user
 
-# ...
-@dataclass
-class User(UserMixin):
-    id: str = ""
-    first_name: str = ""
-    last_name: str = ""
-    email: str = ""
-    password_hash: str = ""
-    oauth_source: str = "self"
-    included_data_sources: str = ""
-    excluded_data_sources: str = ""
-
-    def __str__(self):
-        return "{} {}".format(self.first_name, self.last_name)
-
-    def __repr__(self):
-        return "{} {}".format(self.first_name, self.last_name)
-
-    def set_password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def check_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-    def get_id(self):
-        return self.id
-
-# endregion
-
-
-# region FORMS
-
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, BooleanField, SubmitField, widgets
-from wtforms.validators import (
-    ValidationError,
-    DataRequired,
-    Email,
-    EqualTo,
-    StopValidation,
-)
-from wtforms.fields import SelectMultipleField
-
-
-    # def validate_username(self, username):
-    #     user = db.session.scalar(sa.select(User).where(
-    #         User.username == username.data))
-    #     if user is not None:
-    #         raise ValidationError('Please use a different username.')
-
-    # def validate_email(self, email):
-    #     user = db.session.scalar(sa.select(User).where(
-    #         User.email == email.data))
-    #     if user is not None:
-    #         raise ValidationError('Please use a different email address.')
-
-
-class ProfileForm(FlaskForm):
-    first_name = StringField("First Name", validators=[DataRequired()])
-    last_name = StringField("Last Name", validators=[DataRequired()])
-    email = StringField(
-        "Email",
-        render_kw={"disabled": "disabled"},
-    )
-    submit = SubmitField("Save")
-
-
-class MultiCheckboxField(SelectMultipleField):
-    widget = widgets.ListWidget(html_tag="ol", prefix_label=False)
-    option_widget = widgets.CheckboxInput()
-
-
-class MultiCheckboxAtLeastOne:
-    def __init__(self, message=None):
-        if not message:
-            message = "At least one option must be selected."
-        self.message = message
-
-    def __call__(self, form, field):
-        if len(field.data) == 0:
-            raise StopValidation(self.message)
-
-
-class PreferencesForm(FlaskForm):
-    data_sources = MultiCheckboxField(
-        "Data Sources", validators=[MultiCheckboxAtLeastOne()], coerce=str
-    )
-    submit = SubmitField("Save")
-
-
-# endregion
-
 # region ROUTES
-
-@app.route("/profile", methods=["GET", "POST"])
-@limiter.limit("10 per minute")
-@login_required
-def profile():
-    form = ProfileForm()
-    if form.validate_on_submit():
-        current_user.first_name = form.first_name.data
-        current_user.last_name = form.last_name.data
-        # current_user.email = form.email.data
-        utils.update_user(current_user)
-        flash("Your changes have been saved.", "success")
-        return redirect(url_for("profile"))
-    elif request.method == "GET":
-        form.first_name.data = current_user.first_name
-        form.last_name.data = current_user.last_name
-        form.email.data = current_user.email
-    return render_template(
-        "profile.html",
-        title="Profile",
-        form=form,
-        back_url=session.get("back-url", url_for("public.index")),
-    )
-
-
-@app.route("/preferences", methods=["GET", "POST"])
-@limiter.limit("10 per minute")
-@login_required
-def preferences():
-    form = PreferencesForm()
-    # populate the forms dynamically with the values in the configuration and database
-    data_sources_list = app.config["DATA_SOURCES"]
-    form.data_sources.choices = sorted(
-        [(source, source) for source in app.config["DATA_SOURCES"].keys()]
-    )
-    # if it's a post request and we validated successfully
-    if request.method == "POST" and form.validate_on_submit():
-        # get our choices again, could technically cache these in a list if we wanted but w/e
-        all_sources = app.config["DATA_SOURCES"].keys()
-        # need a list to hold the selections
-        included_data_sources = []
-        excluded_data_sources = []
-        # looping through the choices, we check the choice ID against what was passed in the form
-        for source in all_sources:
-            # when we find a match, we then append the Choice object to our list
-            if source in form.data_sources.data:
-                included_data_sources.append(source)
-            else:
-                excluded_data_sources.append(source)
-        # now all we have to do is update the users choices records
-        current_user.included_data_sources = "; ".join(included_data_sources)
-        current_user.excluded_data_sources = "; ".join(excluded_data_sources)
-        utils.update_user_preferences_data_sources(current_user)
-        flash("Your preferences have been saved.", "success")
-    else:
-        # tell the form what's already selected
-        form.data_sources.data = [
-            source for source in current_user.included_data_sources.split("; ")
-        ]
-    return render_template(
-        "preferences.html",
-        title="Preferences",
-        form=form,
-        back_url=session.get("back-url", url_for("public.index")),
-    )
-
 
 @app.route("/are-embeddings-generated", methods=["GET"])
 @utils.timeit
@@ -796,81 +637,6 @@ def resource_details(source_name, source_id, doi, ts):
     return response
 
 
-# @app.route('/generate-researcher-banner/<string:orcid>', methods=['GET'])
-# @utils.handle_exceptions
-# def generate_researcher_banner(orcid):
-#     generated_banner = generate_researcher_banner(session['researcher:'+orcid])
-#     if generated_banner == "":
-#         import base64
-#         with open('static/images/researcher-default-banner.jpg', "rb") as fh:
-#             generated_banner = base64.b64encode(fh.read()).decode()
-#     return jsonify(generated_banner = f'data:image/jpeg;base64,{generated_banner}')
-
-
-# @app.route('/organization-details/<string:organization_id>/<string:organization_name>', methods=['GET'])
-# @utils.timeit
-# @utils.set_cookies
-# def organization_details(organization_id, organization_name):
-#     try:
-
-#         # Create a response object
-#         """ response = make_response()
-
-#         # Set search-session cookie to the session cookie value of the first visit
-#         if request.cookies.get('search-session') is None:
-#             if request.cookies.get('session') is None:
-#                 response.set_cookie('search-session', str(uuid.uuid4()))
-#             else:
-#                 response.set_cookie('search-session', request.cookies['session'])"""
-
-#         # Call the org_details function from the gepris module to fetch organization details by id
-#         organization, sub_organization, sub_project = org_details(organization_id, organization_name)
-
-#         if organization or sub_organization or sub_project:
-#             # Render the organization-details.html template
-#             return render_template('organization-details.html', organization=organization,
-#                                    sub_organization=sub_organization, sub_project=sub_project)
-#         else:
-#             # Handle the case where organization details are not found (e.g., return a 404 page)
-#             return render_template('error.html', error_message='Organization details not found.')
-
-#     except ValueError as ve:
-#         return render_template('error.html', error_message=str(ve))
-#     except Exception as e:
-#         return render_template('error.html', error_message='An error occurred: ' + str(e))
-
-
-# @app.route('/events-details')
-# @utils.timeit
-# @utils.set_cookies
-# def events_details():
-#     response = make_response(render_template('events-details.html'))
-
-#     # Set search-session cookie to the session cookie value of the first visit
-#     if request.cookies.get('search-session') is None:
-#         if request.cookies.get('session') is None:
-#             response.set_cookie('search-session', str(uuid.uuid4()))
-#         else:
-#             response.set_cookie('search-session', request.cookies['session'])
-
-#     return response
-
-
-# @app.route('/project-details')
-# @utils.timeit
-# @utils.set_cookies
-# def project_details():
-#     response = make_response(render_template('project-details.html'))
-#     # Set search-session cookie to the session cookie value of the first visit
-#     if request.cookies.get('search-session') is None:
-#         if request.cookies.get('session') is None:
-#             response.set_cookie('search-session', str(uuid.uuid4()))
-#         else:
-#             response.set_cookie('search-session', request.cookies['session'])
-
-#     return response
-
-
 @app.route("/digital-obj-details/<path:identifier_with_type>", methods=["GET"])
 @limiter.limit("10 per minute")
 @utils.timeit
@@ -1132,74 +898,6 @@ def merge_objects(object_list, object_type):
     merged_object.source = list(sources)
 
     return merged_object
-
-
-# endregion
-
-
-#region IP BAN
-
-# @app.route('/get-block-list', methods=['GET'])
-# @utils.timeit
-# def get_block_list():
-#     utils.log_activity("get_block_list")
-#     s = '<html><body>'
-#     s += '<table class="table" style="width: 100%"><thead>\n'
-#     s += '<tr><th>ip</th><th>count</th><th>permanent</th><th>url</th><th>timestamp</th></tr>\n'
-#     s += '</thead><tbody>\n'
-#     for k, r in ip_ban.get_block_list().items():
-#         s += '<tr><td>{}</td><td>{}</td><td>{}</td><td>{}</td><td>{}</td></tr>\n'.format(k, r['count'],
-#                                                                                         r.get('permanent', ''),
-#                                                                                         r.get('url', ''),
-#                                                                                         r['timestamp'])
-
-#     s += '</body></html>'
-#     return s
-
-# @app.route('/ip-ban-block/<string:ip_address>', methods=['GET'])
-# @utils.timeit
-# def ip_ban_block(ip_address):
-#     utils.log_activity(f"ip_ban_block: {ip_address}")
-#     ip_ban.block(ip_list=[ip_address])
-#     return f"IP address {ip_address} has been blocked."
-
-# @app.route('/ip-ban-add/<string:ip_address>', methods=['GET'])
-# @utils.timeit
-# def ip_ban_add(ip_address):
-#     utils.log_activity(f"ip_ban_add: {ip_address}")
-#     ip_ban.add(ip=ip_address)
-#     return f"IP address {ip_address} has been added to the list."
-
-# @app.route('/ip-ban-remove/<string:ip_address>', methods=['GET'])
-# @utils.timeit
-# def ip_ban_remove(ip_address):
-#     utils.log_activity(f"ip_ban_remove: {ip_address}")
-#     ip_ban.remove(ip=ip_address)
-#     return f"IP address {ip_address} has been removed from the list."
-
-# @app.route('/ip-whitelist-add/<string:ip_address>', methods=['GET'])
-# @utils.timeit
-# def ip_whitelist_add(ip_address):
-#     utils.log_activity(f"ip_whitelist_add: {ip_address}")
-#     ip_ban.ip_whitelist_add(ip=ip_address)
-#     return f"IP address {ip_address} has been added to the whitelist."
-
-# @app.route('/ip-whitelist-remove/<string:ip_address>', methods=['GET'])
-# @utils.timeit
-# def ip_whitelist_remove(ip_address):
-#     utils.log_activity(f"ip_whitelist_remove: {ip_address}")
-#     ip_ban.ip_whitelist_remove(ip=ip_address)
-#     return f"IP address {ip_address} has been removed from the whitelist."
-
-#endregion
-
-
-@limiter.request_filter
-def ip_whitelist():
-    return request.remote_addr == "127.0.0.1"
-
-
-# region Control Panel
 
 
 # endregion
