+++ nfdi_search_engine/web/account/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("account", __name__)
+
+from . import routes
